Drag touch now mostly implemented, still some work to do on layout, margin and sizing. Drag shadow now entire card stack instead of just the one clicked on.
Split card image view into another class.

Changed CardImageView class to PlayingCardView now extending linear layout instead of imageView this is to allow for the whole stack of cards to be used as the dragshadow.

Created a CardShadowBuilder class extending DragShadowBuilder to allow the shadow to be positioned at the touch point on the view rather than the center of the view.